+++ src/application/README.md
@@ -1,7 +1,9 @@
 # Application
 
-Cette partie est composé des éléments centraux à une application:
-- les popins
-- la barre d'application ( header de taille variable)
-- le centre de message ( erreurs, infos, warning...)
-- les popins de confirmation
+This folder has been deprecated, and is conserved for legacy purpose.
+
+* ErrorCenter can be found in [components/layout/error-center](../components/layout/error)
+* Popin can be found in [components/popin](../components/popin)
+* ConfirmationPopin can be found in [components/confirm/popin](../components/confirm/popin)
+
+Any other old component should have an equivalent in [components/layout](../components/layout)
