+++ src/application/index.js
@@ -1,13 +1,10 @@
 import popin from './popin';
-import confirmationPopin from './confirmation-popin';
+import confirmationPopin from '../components/confirm/popin';
 import errorCenter from './error-center';
-import loadingBar from './loading-bar';
 
 export {
-    popin, confirmationPopin,
-    errorCenter, loadingBar
+    popin, confirmationPopin, errorCenter
 };
 export default {
-    popin, confirmationPopin,
-    errorCenter, loadingBar
+    popin, confirmationPopin, errorCenter
 };
