+++ src/common/README.md
@@ -1,0 +1,6 @@
+# Common
+
+This folder has been deprecated, and is conserved for legacy purpose.
+
+* Old components in this folder have been deleted (equivalent can be found in [components/layout](../components/layout))
+* Mixin have been moved to [mixin](../mixin)
