+++ src/components/confirm/style/confirmation-popin.scss
@@ -1,0 +1,13 @@
+[data-focus='confirmation-popin'] {
+    [data-focus='button-stack'] {
+      text-align: center;
+      position: absolute;
+      bottom: 0;
+      width: 100%;
+  
+      button {
+        margin: 20px;
+      }
+    }
+  }
+
