+++ src/components/layout/error-center/style/error-center.scss
